requirements.txt
coincurve>=20.0.0
pysha3>=1.0.2
pytest>=7.0.0
